(folders): reorganize files to simplify exports
