+++ src/components/atomics/index.ts
@@ -1,0 +1,5 @@
+import AButton from './AButton.vue'
+import AIcon from './AIcon.vue'
+import ASwitch from './ASwitch.vue'
+
+export { AButton, AIcon, ASwitch }
