+++ src/components/inputs/index.ts
@@ -1,0 +1,5 @@
+import ACheckbox from './ACheckbox.vue'
+import AInput from './AInput.vue'
+import AInputErrors from './AInputErrors.vue'
+
+export { ACheckbox, AInput, AInputErrors }
