+++ src/constants/index.ts
@@ -1,0 +1,6 @@
+import * as AliasColors from './AliasColors'
+import * as AliasIcons from './AliasIcons'
+import * as AliasSizes from './AliasSizes'
+import * as HtmlTypes from './HtmlTypes'
+
+export { AliasColors, AliasIcons, AliasSizes, HtmlTypes }
