+++ src/index.ts
@@ -1,25 +1,9 @@
 import 'tailwindcss/tailwind.css'
 
-import ASwitch from './components/atomics/ASwitch.vue'
-import ACheckbox from './components/inputs/ACheckbox.vue'
-import AButton from './components/atomics/AButton.vue'
-import AIcon from './components/atomics/AIcon.vue'
-import AInput from './components/inputs/AInput.vue'
-import { AliasColorNames } from './AliasColors'
-import * as AliasSizes from './AliasSizes'
-import * as HtmlTypes from './HtmlTypes'
-import * as UseValidation from './composables/useValidation'
-import * as AliasValidations from '../utils/validation'
+import * as Atomics from './components/atomics'
+import * as Inputs from './components/inputs'
+import * as Composables from './composables'
+import * as AliasConstants from './constants'
+import * as AliasUtils from './utils'
 
-export {
-  AButton,
-  AIcon,
-  AInput,
-  ACheckbox,
-  ASwitch,
-  AliasColorNames,
-  AliasSizes,
-  HtmlTypes,
-  UseValidation,
-  AliasValidations,
-}
+export { Atomics, Inputs, Composables, AliasConstants, AliasUtils }
